form_action.asp (panyuxin.com)
Form_action.asp is an important file when you want to collect
information from the form that is located on your website. It processes
the value that submitted by the visitors of your website and stores it.
Many friends of mine do not know how to programme this file so here I
publish the one on my website (panyuxin.com) to share my opinion.

{www.panyuxin.com}